Router structure change
