+++ src/app/(auth)/layout.tsx
@@ -1,8 +1,7 @@
 import BrandLogo from "@/components/ui/BrandLogo"
 import Link from "next/link"
-import RegisterForm from "@/components/RegisterForm"
 
-export default function RegisterPage() {
+export default function AuthLayout({ children }: { children: React.ReactNode }) {
   return (
     <div className="flex flex-col lg:flex-row grow">
       <div className="p-4 flex flex-1">
@@ -10,15 +9,7 @@ export default function RegisterPage() {
           <Link href="/" className="flex justify-center lg:justify-normal">
             <BrandLogo size="big" />
           </Link>
-          <div className="text-center lg:text-left leading-10 lg:leading-[60px]">
-            <h1 className="text-4xl lg:text-5xl font-display font-semibold">
-              Stwórz nowe konto
-            </h1>
-            <h2 className="text-lg lg:text-xl text-muted">
-              Zarejestruj się za pomocą adresu email
-            </h2>
-          </div>
-          <RegisterForm />
+          {children}
           <p className="text-muted text-center">
             Kontynuując zgadzasz się na warunki
             <br />
