--- src/app/(special)/layout.tsx
@@ -1,3 +1,0 @@
-export default function SpecialLayout({ children }: { children: React.ReactNode }) {
-  return children
-}
