+++ src/app/providers.tsx
@@ -1,7 +1,7 @@
 "use client"
 
 import { ThemeProvider } from "next-themes"
-import { AuthContext } from "@/components/AuthContext"
+import { AuthContext } from "@/components/auth/AuthContext"
 import { useEffect, useState } from "react"
 import { createClient } from "@/lib/supabase/client"
 import { User } from "@supabase/supabase-js"
