+++ src/components/auth/RegisterForm.tsx
@@ -3,11 +3,10 @@
 import { useForm } from "react-hook-form"
 import { zodResolver } from "@hookform/resolvers/zod"
 import * as z from "zod"
-import { Input } from "./ui/Input"
+import { Input } from "../ui/Input"
 import React from "react"
-import { login } from "@/app/(special)/login/actions"
+import { signup } from "@/app/(auth)/actions"
 import { toast } from "sonner"
-import { createClient } from "@/lib/supabase/client"
 
 const schema = z.object({
   email: z
@@ -17,7 +16,7 @@ const schema = z.object({
   password: z.string().min(8, { message: "Hasło musi się składać z minimum 8 znaków" }),
 })
 
-export default function LoginForm() {
+export default function RegisterForm() {
   const {
     register,
     handleSubmit,
@@ -28,14 +27,14 @@ export default function LoginForm() {
   return (
     <form
       onSubmit={handleSubmit(async (data) => {
-        const error = await login(data)
-        if (error && error.status === 400) {
-          toast.error("Błędne dane logowania")
+        const error = await signup(data)
+        if (error) {
+          toast.error("W trakcie rejestracji wystąpił błąd")
           return
         }
-        toast.success("Zalogowano")
+        toast.success("Zarejestrowano")
       })}
-      className="flex flex-col gap-4"
+      className="w-full flex flex-col gap-4"
     >
       <div className="flex flex-col gap-2">
         <label htmlFor="email">Email</label>
@@ -52,7 +51,7 @@ export default function LoginForm() {
         )}
         <Input
           type="submit"
-          value="Zaloguj się"
+          value="Stwórz konto"
           className="bg-primary hover:cursor-pointer"
         />
       </div>
